fix remaining tests and eliminate the photo picker abstraction as it doesnt work well in the UI test architecture
