+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackWidget.swift
@@ -19,12 +19,6 @@ struct SentryUserFeedbackWidget {
             
             lazy var button = SentryUserFeedbackWidgetButtonView(config: config, action: { _ in
                 self.setWidget(visible: false)
-                #if SENTRY_TEST || SENTRY_TEST_CI
-                // the iOS photo picker UI doesn't play nicely with XCUITest, so we need to mock it
-//                if let sdkEnv = ProcessInfo.processInfo.environment["--io.sentry.sdk-environment"], sdkEnv == "ui-tests" {
-                    config.formConfig.photoPicker = TestSentryPhotoPicker()
-//                }
-                #endif // SENTRY_TEST || SENTRY_TEST_CI
                 let form = SentryUserFeedbackFormController(config: self.config, delegate: self)
                 form.presentationController?.delegate = self
                 self.present(form, animated: self.config.animations) {
